refactor(pg): Move common components to separate file
